+++ src/Configuration/Kitten.js
