+++ src/Configuration/Redux.js
@@ -1,0 +1,6 @@
+import { createStore as create } from 'redux'
+import RootReducer from '../Redux/Reducers'
+
+const initialState = {}
+
+export const createStore = () => create(RootReducer, initialState)
