+++ src/Configuration/Screens.js
@@ -1,12 +1,16 @@
 import { Navigation } from 'react-native-navigation'
 import { Provider } from 'react-redux'
 
-import SignInScreen from './Containers/SignInScreen'
-import RideList from './Containers/RideList'
-import TestScreen from './Containers/TestScreen'
+import SignInScreen from '../Modules/Authentication/Containers/SignInScreen'
+import RideList from '../Modules/RideRequest/Containers/RideList'
+import TestScreen from '../Modules/Shared/Containers/TestScreen'
 
-export default function (store) {
+const registerScreens = (store) => {
   Navigation.registerComponent('carona.signIn', () => SignInScreen, store, Provider)
   Navigation.registerComponent('carona.rideList', () => RideList, store, Provider)
   Navigation.registerComponent('carona.testScreen', () => TestScreen, store, Provider)
+}
+
+export default {
+  registerScreens
 }
