+++ src/Modules/Shared/Components/Ride.js
