+++ src/Redux/Types/TestActionTypes.js
@@ -1,0 +1,1 @@
+export const TEST = 'test'
