+++ src/Redux/Types/index.js
@@ -1,1 +1,1 @@
-export const TEST = 'test'
+export * from './TestActionTypes'
